+++ arviz/tests/base_tests/test_diagnostics.py
@@ -1,28 +1,28 @@
 """Test Diagnostic methods"""
 # pylint: disable=redefined-outer-name, no-member, too-many-public-methods
 import os
+
 import numpy as np
-from numpy.testing import assert_almost_equal, assert_array_almost_equal
 import pandas as pd
 import pytest
+from numpy.testing import assert_almost_equal, assert_array_almost_equal
 
-from ...data import load_arviz_data, from_cmdstan
+from ...data import from_cmdstan, load_arviz_data
 from ...plots.plot_utils import xarray_var_iter
-from ...stats import bfmi, rhat, ess, mcse, geweke
+from ...rcparams import rcParams
+from ...stats import bfmi, ess, geweke, mcse, rhat
 from ...stats.diagnostics import (
-    ks_summary,
+    _conv_quantile,
     _ess,
     _ess_quantile,
-    _multichain_statistics,
     _mc_error,
+    _multichain_statistics,
     _rhat,
     _rhat_rank,
-    _z_scale,
-    _conv_quantile,
     _split_chains,
+    _z_scale,
+    ks_summary,
 )
-from ...utils import Numba
-from ...rcparams import rcParams
 
 # For tests only, recommended value should be closer to 1.01-1.05
 # See discussion in https://github.com/stan-dev/rstan/pull/618
@@ -536,85 +536,3 @@ class TestDiagnostics:
         if chains is None:
             chains = 1
         assert split_data.shape == (chains * 2, draws // 2)
-
-
-def test_numba_bfmi():
-    """Numba test for bfmi."""
-    state = Numba.numba_flag
-    school = load_arviz_data("centered_eight")
-    data_md = np.random.rand(100, 100, 10)
-    Numba.disable_numba()
-    non_numba = bfmi(school.posterior["mu"].values)
-    non_numba_md = bfmi(data_md)
-    Numba.enable_numba()
-    with_numba = bfmi(school.posterior["mu"].values)
-    with_numba_md = bfmi(data_md)
-    assert np.allclose(non_numba_md, with_numba_md)
-    assert np.allclose(with_numba, non_numba)
-    assert state == Numba.numba_flag
-
-
-@pytest.mark.parametrize("method", ("rank", "split", "folded", "z_scale", "identity"))
-def test_numba_rhat(method):
-    """Numba test for mcse."""
-    state = Numba.numba_flag
-    school = np.random.rand(100, 100)
-    Numba.disable_numba()
-    non_numba = rhat(school, method=method)
-    Numba.enable_numba()
-    with_numba = rhat(school, method=method)
-    assert np.allclose(with_numba, non_numba)
-    assert Numba.numba_flag == state
-
-
-@pytest.mark.parametrize("method", ("mean", "sd", "quantile"))
-def test_numba_mcse(method, prob=None):
-    """Numba test for mcse."""
-    state = Numba.numba_flag
-    school = np.random.rand(100, 100)
-    if method == "quantile":
-        prob = 0.80
-    Numba.disable_numba()
-    non_numba = mcse(school, method=method, prob=prob)
-    Numba.enable_numba()
-    with_numba = mcse(school, method=method, prob=prob)
-    assert np.allclose(with_numba, non_numba)
-    assert Numba.numba_flag == state
-
-
-def test_ks_summary_numba():
-    """Numba test for ks_summary."""
-    state = Numba.numba_flag
-    data = np.random.randn(100, 100)
-    Numba.disable_numba()
-    non_numba = (ks_summary(data)["Count"]).values
-    Numba.enable_numba()
-    with_numba = (ks_summary(data)["Count"]).values
-    assert np.allclose(non_numba, with_numba)
-    assert Numba.numba_flag == state
-
-
-def test_geweke_numba():
-    """Numba test for geweke."""
-    state = Numba.numba_flag
-    data = np.random.randn(100)
-    Numba.disable_numba()
-    non_numba = geweke(data)
-    Numba.enable_numba()
-    with_numba = geweke(data)
-    assert np.allclose(non_numba, with_numba)
-    assert Numba.numba_flag == state
-
-
-@pytest.mark.parametrize("batches", (1, 20))
-@pytest.mark.parametrize("circular", (True, False))
-def test_mcse_error_numba(batches, circular):
-    """Numba test for mcse_error."""
-    data = np.random.randn(100, 100)
-    state = Numba.numba_flag
-    Numba.disable_numba()
-    non_numba = _mc_error(data, batches=batches, circular=circular)
-    Numba.enable_numba()
-    with_numba = _mc_error(data, batches=batches, circular=circular)
-    assert np.allclose(non_numba, with_numba)
-    assert state == Numba.numba_flag